refactor: 배틀 리다이렉트 로직을 BattleGuard로 분리 (#225)
- 리다이렉트 로직을 BattleGuard로 이동
- App.tsx에서 리다이렉트 로직 제거
